packages/speech-to-text/speech-to-text/requirements.txt

google-api-core
google-auth
google-cloud-core
google-cloud-speech
google-cloud-translate
PyAudio

